Basic refactor + loading books from user object + api request with own http service
